pyproject.toml
[project]
name = "gamelens-collector-service"
version = "0.1.0"
description = "GameLens Data Collector Service."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gevent>=24.2.1",
    "gevent-websocket>=0.10.1",
    "flasgger>=0.9.7.1",
    "flask>=3.1.2",
    "flask-socketio>=5.6.0",
    "gunicorn>=25.1.0",
    "psycopg[binary,pool]>=3.3.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
]
